pyproject.toml
[project]
name = "echo-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.11",
    "langchain-aws>=0.2.30",
    "mcp-use>=1.3.9",
    "numpy>=2.3.2",
    "python-dotenv>=1.0.0",
    "zstandard>=0.23.0",
]
